// client/src/generated/torii.ts
export interface SubscriptionRequest {
    /**
     * Unique client ID - used to track client connection
     *
     * @generated from protobuf field: string client_id = 1
     */
    client_id: string;
    /**
     * Topics to subscribe to (will replace existing subscriptions)
     *
     * @generated from protobuf field: repeated torii.TopicSubscription topics = 2
     */
    topics: TopicSubscription[];
    /**
     * Topics to unsubscribe from (by topic name)
     *
     * @generated from protobuf field: repeated string unsubscribe_topics = 3
     */
    unsubscribe_topics: string[];
}
/**
 * Topic subscription with optional filters
 *
 * @generated from protobuf message torii.TopicSubscription
 */
export interface TopicSubscription {
    /**
     * Topic name (e.g., "sql", "logs")
     *
     * @generated from protobuf field: string topic = 1
     */
    topic: string;
    /**
     * Optional filters for this topic (generic key-value)
     * Empty map = subscribe to all updates in this topic
     *
     * @generated from protobuf field: map<string, string> filters = 2
     */
    filters: {
        [key: string]: string;
    };
    /**
     * Optional: Sink-specific structured filter (protobuf Any)
     * Allows sinks to define rich filter schemas
     *
     * @generated from protobuf field: google.protobuf.Any filter_data = 3
     */
    filter_data?: Any;
}
/**
 * Update sent to subscribed clients
 *
 * @generated from protobuf message torii.TopicUpdate
 */
export interface TopicUpdate {
    /**
     * Topic this update belongs to
     *
     * @generated from protobuf field: string topic = 1
     */
    topic: string;
    /**
     * Type of update
     *
     * @generated from protobuf field: torii.UpdateType update_type = 2
     */
    update_type: UpdateType;
    /**
     * Timestamp of the update
     *
     * @generated from protobuf field: int64 timestamp = 3
     */
    timestamp: bigint;
    /**
     * Type identifier for the data (e.g., "sql.row_inserted", "log.entry")
     * Used by client to determine how to deserialize data
     *
     * @generated from protobuf field: string type_id = 4
     */
    type_id: string;
    /**
     * Sink-specific structured data (protobuf Any)
     * Contains the actual update data from the sink
     *
     * @generated from protobuf field: google.protobuf.Any data = 5
     */
    data?: Any;
}
/**
 * @generated from protobuf enum torii.UpdateType
 */
export enum UpdateType {
    /**
     * @generated from protobuf enum value: CREATED = 0;
     */
    CREATED = 0,
    /**
     * @generated from protobuf enum value: UPDATED = 1;
     */
    UPDATED = 1,
    /**
     * @generated from protobuf enum value: DELETED = 2;
     */
    DELETED = 2
}
// ...
class SubscriptionRequest$Type extends MessageType<SubscriptionRequest> {
    constructor() {
        super("torii.SubscriptionRequest", [
            { no: 1, name: "client_id", kind: "scalar", localName: "client_id", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "topics", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => TopicSubscription },
            { no: 3, name: "unsubscribe_topics", kind: "scalar", localName: "unsubscribe_topics", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<SubscriptionRequest>): SubscriptionRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.client_id = "";
        message.topics = [];
        message.unsubscribe_topics = [];
        if (value !== undefined)
            reflectionMergePartial<SubscriptionRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SubscriptionRequest): SubscriptionRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string client_id */ 1:
                    message.client_id = reader.string();
                    break;
                case /* repeated torii.TopicSubscription topics */ 2:
                    message.topics.push(TopicSubscription.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* repeated string unsubscribe_topics */ 3:
                    message.unsubscribe_topics.push(reader.string());
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SubscriptionRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string client_id = 1; */
        if (message.client_id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.client_id);
        /* repeated torii.TopicSubscription topics = 2; */
        for (let i = 0; i < message.topics.length; i++)
            TopicSubscription.internalBinaryWrite(message.topics[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* repeated string unsubscribe_topics = 3; */
        for (let i = 0; i < message.unsubscribe_topics.length; i++)
            writer.tag(3, WireType.LengthDelimited).string(message.unsubscribe_topics[i]);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message torii.SubscriptionRequest
 */
export const SubscriptionRequest = new SubscriptionRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class TopicSubscription$Type extends MessageType<TopicSubscription> {
    constructor() {
        super("torii.TopicSubscription", [
            { no: 1, name: "topic", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "filters", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "scalar", T: 9 /*ScalarType.STRING*/ } },
            { no: 3, name: "filter_data", kind: "message", localName: "filter_data", T: () => Any }
        ]);
    }
    create(value?: PartialMessage<TopicSubscription>): TopicSubscription {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.topic = "";
        message.filters = {};
        if (value !== undefined)
            reflectionMergePartial<TopicSubscription>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: TopicSubscription): TopicSubscription {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string topic */ 1:
                    message.topic = reader.string();
                    break;
                case /* map<string, string> filters */ 2:
                    this.binaryReadMap2(message.filters, reader, options);
                    break;
                case /* google.protobuf.Any filter_data */ 3:
                    message.filter_data = Any.internalBinaryRead(reader, reader.uint32(), options, message.filter_data);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    private binaryReadMap2(map: TopicSubscription["filters"], reader: IBinaryReader, options: BinaryReadOptions): void {
        let len = reader.uint32(), end = reader.pos + len, key: keyof TopicSubscription["filters"] | undefined, val: TopicSubscription["filters"][any] | undefined;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case 1:
                    key = reader.string();
                    break;
                case 2:
                    val = reader.string();
                    break;
                default: throw new globalThis.Error("unknown map entry field for torii.TopicSubscription.filters");
            }
        }
        map[key ?? ""] = val ?? "";
    }
    internalBinaryWrite(message: TopicSubscription, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string topic = 1; */
        if (message.topic !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.topic);
        /* map<string, string> filters = 2; */
        for (let k of globalThis.Object.keys(message.filters))
            writer.tag(2, WireType.LengthDelimited).fork().tag(1, WireType.LengthDelimited).string(k).tag(2, WireType.LengthDelimited).string(message.filters[k]).join();
        /* google.protobuf.Any filter_data = 3; */
        if (message.filter_data)
            Any.internalBinaryWrite(message.filter_data, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message torii.TopicSubscription
 */
export const TopicSubscription = new TopicSubscription$Type();
// ...
export const TopicUpdate = new TopicUpdate$Type();